Fixes to pass full review of amo editors
This commit addresses the following issues reported by Mozilla Editors:
    * It's not clear how to use your add-on after it is installed. You should open a first-run page in a new tab explaining to users what they need to do to configure it.

The commit adds a new web page that explains step by step how the user should configure the add-on.

    * This add-on consumes considerable system resources without it being remotely apparent to the user that it is even running. You'll need to show some indicator of its continued operation, ideally in the form of something like an Add-on Bar widget displaying its progress.

The commit adds a new panel and widget notifying the user the poll interval.

    * Please do not import the same module every time you need to use it. Modules should be required once, ideally at the head of the file, and then stored for future use.

The commit has refactored the code, fixing this problem.